+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Added title (hover) texts to degree programme favorite widget (HUB-200)
 * Feedback: User email address as the from address (HUB-190)
 * Display search filters only when more than one type of result (HUB-202)
+* Redirect external active degree programme calls to front page (HUB-184)
 
 
 ## 1.2